+++ flake.nix
@@ -11,13 +11,13 @@
       packages = forAll (pkgs: rec {
         undo = pkgs.stdenv.mkDerivation {
           pname = "undo";
-          version = "0.2.2";
+          version = "0.2.3";
           src = self;
           nativeBuildInputs = [ pkgs.go ];
           # no Go module downloads needed: the CLI is stdlib-only
           buildPhase = ''
             export HOME=$TMPDIR GOCACHE=$TMPDIR/gocache
-            make CC=$CC VERSION=v0.2.2
+            make CC=$CC VERSION=v0.2.3
           '';
           installPhase = ''
             make install PREFIX=$out